+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `jj diff --stat` has been implemented. It shows a histogram of the changes,
   same as `git diff --stat`. Fixes [#2066](https://github.com/martinvonz/jj/issues/2066)
 
+* `jj git fetch --all-remotes` has been implemented. It fetches all remotes
+  instead of just the default remote
+
 ### Breaking changes
 
 * The minimum supported Rust version (MSRV) is now 1.71.0.